#4 Moving tests from handoff to pingpong module
