+++ tests/test_handoff.py
@@ -1,10 +1,8 @@
-from pathlib import Path
 from typing import List
 
 import pandas as pd
 import pytest
-
-from waste import handoff, core
+from waste import handoff
 
 
 @pytest.fixture
@@ -29,50 +27,3 @@ def test_negative_duration(assets_path):
     result = handoff.identify(log_path)
     assert sum(result['duration_sum_seconds'] < 0) == 0
 
-
-def test_ping_pong_identification(assets_path):
-    args = [
-        {
-            'name': 'A',
-            'log_path': assets_path / 'PurchasingExample.xes',
-            'case_path': assets_path / 'ping_pong_case.csv',
-            'expected': [{'source_activity': 'Create Request for Quotation',
-                          'source_resource': 'Kim Passa',
-                          'destination_activity': 'Analyze Request for Quotation',
-                          'destination_resource': 'Karel de Groot'}]
-        },
-        {
-            'name': 'B',
-            'log_path': assets_path / 'PurchasingExample.xes',
-            'case_path': assets_path / 'ping_pong_case_fake.csv',
-            'expected': []
-        },
-        {
-            'name': 'C',
-            'log_path': assets_path / 'PurchasingExample.xes',
-            'case_path': assets_path / 'double_ping_pong_case.csv',
-            'expected': [{'source_activity': 'Create Request for Quotation',
-                          'source_resource': 'Kim Passa',
-                          'destination_activity': 'Analyze Request for Quotation',
-                          'destination_resource': 'Karel de Groot'},
-                         {'source_activity': 'Approve Purchase Order for payment',
-                          'source_resource': 'Karel de Groot',
-                          'destination_activity': 'Send Invoice',
-                          'destination_resource': 'Kiu Kan'}]
-        }
-    ]
-
-    for arg in args:
-        print(f"\rTest case {arg['name']}")
-        log_path: Path = arg['log_path']
-        case_path: Path = arg['case_path']
-        expected = arg['expected']
-
-        log = core.lifecycle_to_interval(log_path)
-        parallel_activities = core.parallel_activities_with_heuristic_oracle(log)
-
-        case = pd.read_csv(case_path)
-        assert case is not None and len(case) > 0
-
-        result = handoff._identify_ping_pongs_per_case(case, parallel_activities)
-        assert result == expected
